# docs/swagger.yaml
basePath: /api
definitions:
...
  api.DisciplineListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Discipline'
        type: array
      meta:
        $ref: '#/definitions/api.PaginationMeta'
    type: object
  api.DisciplineResponse:
    properties:
      data:
        $ref: '#/definitions/models.Discipline'
      meta: {}
    type: object
  api.EmptyResponse:
    properties:
      data: {}
      meta: {}
    type: object
  api.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorDetail'
    type: object
  api.GamePlayerStatListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.GamePlayerStat'
        type: array
      meta:
        $ref: '#/definitions/api.PaginationMeta'
    type: object
  api.GamePlayerStatResponse:
    properties:
      data:
        $ref: '#/definitions/models.GamePlayerStat'
      meta: {}
    type: object
...
  api.PaginationMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  api.disciplineRequest:
    properties:
      code:
        type: string
      description:
        type: string
      icon_url:
        type: string
      is_active:
        type: boolean
      metadata:
        type: object
      name:
        type: string
      team_size:
        type: integer
    required:
    - code
    - name
    type: object
  api.gamePlayerStatRequest:
    properties:
      assists:
        type: integer
      damage_dealt:
        type: integer
      deaths:
        type: integer
      game_id:
        type: integer
      gold_earned:
        type: integer
      hero_name:
        type: string
      kills:
        type: integer
      player_id:
        type: integer
      team_id:
        type: integer
      was_mvp:
        type: boolean
    required:
    - game_id
    - player_id
    type: object
...
  models.Discipline:
    properties:
      code:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      metadata:
        type: object
      name:
        type: string
      team_size:
        type: integer
    type: object
  models.GamePlayerStat:
    properties:
      assists:
        type: integer
      damage_dealt:
        type: integer
      deaths:
        type: integer
      game_id:
        type: integer
      gold_earned:
        type: integer
      hero_name:
        type: string
      id:
        type: integer
      kda_ratio:
        type: number
      kills:
        type: integer
      player_id:
        type: integer
      team_id:
        type: integer
      was_mvp:
        type: boolean
    type: object
...
info:
  contact: {}
  description: REST API for disciplines, teams, tournaments, matches, and reports.
  title: DB Course Project API
  version: "1.0"
paths:
...
  /disciplines:
    get:
      parameters:
      - description: Search by name or code
        in: query
        name: search
        type: string
      - description: Filter by active flag
        in: query
        name: is_active
        type: boolean
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DisciplineListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List disciplines
      tags:
      - Disciplines
    post:
      consumes:
      - application/json
      parameters:
      - description: Discipline payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.disciplineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.DisciplineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create discipline
      tags:
      - Disciplines
  /disciplines/{id}:
    delete:
      parameters:
      - description: Discipline ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/api.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete discipline
      tags:
      - Disciplines
    get:
      parameters:
      - description: Discipline ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DisciplineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get discipline
      tags:
      - Disciplines
    put:
      consumes:
      - application/json
      parameters:
      - description: Discipline ID
        in: path
        name: id
        required: true
        type: integer
      - description: Discipline payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.disciplineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DisciplineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update discipline
      tags:
      - Disciplines
  /game-player-stats:
    get:
      parameters:
      - description: Game ID
        in: query
        name: game_id
        type: integer
      - description: Player ID
        in: query
        name: player_id
        type: integer
      - description: Team ID
        in: query
        name: team_id
        type: integer
      - description: Filter by MVP
        in: query
        name: was_mvp
        type: boolean
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GamePlayerStatListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List game player stats
      tags:
      - GamePlayerStats
    post:
      consumes:
      - application/json
      parameters:
      - description: Game player stats payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.gamePlayerStatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.GamePlayerStatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create game player stats
      tags:
      - GamePlayerStats
  /game-player-stats/{id}:
    delete:
      parameters:
      - description: Stat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/api.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete game player stats
      tags:
      - GamePlayerStats
    get:
      parameters:
      - description: Stat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GamePlayerStatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get game player stats
      tags:
      - GamePlayerStats
    put:
      consumes:
      - application/json
      parameters:
      - description: Stat ID
        in: path
        name: id
        required: true
        type: integer
      - description: Game player stats payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.gamePlayerStatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GamePlayerStatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update game player stats
      tags:
      - GamePlayerStats
...
swagger: "2.0"